+++ R/ES_SE.R
@@ -1,3 +1,6 @@
+#'
+#' @import RPEIF
+#'
 #' @title Standard Error Estimate for Expected Shortfall (ES) of Returns
 #'
 #' @description \code{ES.SE} computes the standard error of the expected shortfall of the returns.
@@ -13,6 +16,7 @@
 #' @param d.GLM.EN Order of the polynomial for the Exponential or Gamma fitting. Default polynomial order of 5.
 #' @param freq.include Frequency domain inclusion criteria. Must be one of "All" (default), "Decimate" or "Truncate."
 #' @param freq.par Percentage of the frequency used if \code{"freq.include"} is "Decimate" or "Truncate." Default is 0.5.
+#' @param corOut Return correlation of the returns or the influence function transformed returns. Must be one of "retCor", "retIFCor" or "none" (default).
 #' @param ... Additional parameters.
 #'
 #' @return A vector or a list depending on \code{se.method}.
@@ -94,6 +98,7 @@ ES.SE <- function(data, p=0.95,
                   se.method=c("IFiid","IFcor","IFcorAdapt","IFcorPW","BOOTiid","BOOTcor")[1:2],
                   cleanOutliers=FALSE, fitting.method=c("Exponential", "Gamma")[1], d.GLM.EN=5,
                   freq.include=c("All", "Decimate", "Truncate")[1], freq.par=0.5,
+                  corOut = c("none", "retCor","retIFCor", "retIFCorPW")[1],
                   ...)
 { # @author Brian G. Peterson and Xin Chen
 
@@ -165,6 +170,11 @@ ES.SE <- function(data, p=0.95,
                                     freq.include=freq.include, freq.par=freq.par,
                                     ...)
       }
+
+      # Adding the correlations to the list
+      res <- Add_Correlations(res=res, data=data, cleanOutliers=cleanOutliers, corOut=corOut, IF.func=IF.ES, ...)
+
+      # Returning the output
       return(res)
     }
   }