+++ R/LPM_SE.R
@@ -1,3 +1,6 @@
+#'
+#' @import RPEIF
+#'
 #' @title Standard Error Estimate for Lower Partial Moment (LPM) of Returns
 #'
 #' @description \code{LPM.SE} computes the standard error of the LPM of the returns.
@@ -14,6 +17,7 @@
 #' @param d.GLM.EN Order of the polynomial for the Exponential or Gamma fitting. Default polynomial order of 5.
 #' @param freq.include Frequency domain inclusion criteria. Must be one of "All" (default), "Decimate" or "Truncate."
 #' @param freq.par Percentage of the frequency used if \code{"freq.include"} is "Decimate" or "Truncate." Default is 0.5.
+#' @param corOut Return correlation of the returns or the influence function transformed returns. Must be one of "retCor", "retIFCor" or "none" (default).
 #' @param ... Additional parameters.
 #'
 #' @return A vector or a list depending on \code{se.method}.
@@ -41,6 +45,7 @@ LPM.SE = function(data, const = 0, k = 1,
                   se.method=c("IFiid","IFcor","IFcorAdapt","IFcorPW","BOOTiid","BOOTcor")[1:2],
                   cleanOutliers=FALSE, fitting.method=c("Exponential", "Gamma")[1], d.GLM.EN=5,
                   freq.include=c("All", "Decimate", "Truncate")[1], freq.par=0.5,
+                  corOut = c("none", "retCor","retIFCor", "retIFCorPW")[1],
                   ...){
   data = checkData(data)
   myLPM = t(apply(data, 2, LPM, const = const, k = k, ...))
@@ -58,6 +63,11 @@ LPM.SE = function(data, const = 0, k = 1,
                                   freq.include=freq.include, freq.par=freq.par,
                                   ...)
     }
+
+    # Adding the correlations to the list
+    res <- Add_Correlations(res=res, data=data, cleanOutliers=cleanOutliers, corOut=corOut, IF.func=IF.LPM, ...)
+
+    # Returning the output
     return(res)
   }
 }