+++ R/Measures.R
@@ -24,7 +24,7 @@ RES=function(data,...,alpha=0.01,beta=0.1){
 }
 
 # Compute sample Sortino Ratio with mean threshold
-SoR = function(data, ..., rf = 0, MAR = 0, threshold=c("mean", "const")[1]){
+SoR = function(data, rf = 0, MAR = 0, threshold=c("mean", "const")[1], ...){
 
   if(threshold=="mean"){
     mu.hat = mean(data)