+++ R/SemiSD_SE.R
@@ -1,3 +1,6 @@
+#'
+#' @import RPEIF
+#'
 #' @title Standard Error Estimate for Semi-Standared Deviation (SemiSD) of Returns
 #'
 #' @description \code{SemiSD.SE} computes the standard error of the SSD of the returns.
@@ -13,6 +16,7 @@
 #' @param d.GLM.EN Order of the polynomial for the Exponential or Gamma fitting. Default polynomial order of 5.
 #' @param freq.include Frequency domain inclusion criteria. Must be one of "All" (default), "Decimate" or "Truncate."
 #' @param freq.par Percentage of the frequency used if \code{"freq.include"} is "Decimate" or "Truncate." Default is 0.5.
+#' @param corOut Return correlation of the returns or the influence function transformed returns. Must be one of "retCor", "retIFCor" or "none" (default).
 #' @param ... Additional parameters.
 #'
 #' @return A vector or a list depending on \code{se.method}.
@@ -39,6 +43,7 @@ SemiSD.SE <- function(data, rf=0,
                       se.method=c("IFiid","IFcor","IFcorAdapt","IFcorPW","BOOTiid","BOOTcor")[1:2],
                       cleanOutliers=FALSE, fitting.method=c("Exponential", "Gamma")[1], d.GLM.EN = 5,
                       freq.include=c("All", "Decimate", "Truncate")[1], freq.par=0.5,
+                      corOut = c("none", "retCor","retIFCor", "retIFCorPW")[1],
                       ...){
   data = checkData(data)
   mySSD = t(apply(data, 2, SemiSD, rf, ...))
@@ -57,6 +62,11 @@ SemiSD.SE <- function(data, rf=0,
                                   freq.include=freq.include, freq.par=freq.par,
                                   ...)
     }
+
+    # Adding the correlations to the list
+    res <- Add_Correlations(res=res, data=data, cleanOutliers=cleanOutliers, corOut=corOut, IF.func=IF.SemiSD, ...)
+
+    # Returning the output
     return(res)
   }
 }