+++ R/StdDev_SE.R
@@ -1,3 +1,6 @@
+#'
+#' @import RPEIF
+#'
 #' @title Standard Error Estimate for Standard Deviation (StdDev) of Returns
 #'
 #' @description \code{StdDev.SE} computes the standard error of the standard deviation of the returns.
@@ -12,6 +15,7 @@
 #' @param d.GLM.EN Order of the polynomial for the Exponential or Gamma fitting. Default polynomial order of 5.
 #' @param freq.include Frequency domain inclusion criteria. Must be one of "All" (default), "Decimate" or "Truncate."
 #' @param freq.par Percentage of the frequency used if \code{"freq.include"} is "Decimate" or "Truncate." Default is 0.5.
+#' @param corOut Return correlation of the returns or the influence function transformed returns. Must be one of "retCor", "retIFCor" or "none" (default).
 #' @param ... Additional parameters.
 #'
 #' @return A vector or a list depending on \code{se.method}.
@@ -39,6 +43,7 @@ StdDev.SE <- function(data,
                       se.method=c("IFiid","IFcor","IFcorAdapt","IFcorPW","BOOTiid","BOOTcor")[1:2],
                       cleanOutliers=FALSE, fitting.method=c("Exponential", "Gamma")[1], d.GLM.EN = 5,
                       freq.include=c("All", "Decimate", "Truncate")[1], freq.par=0.5,
+                      corOut = c("none", "retCor","retIFCor", "retIFCorPW")[1],
                       ...){
 
   # Forcing the following parameters
@@ -72,9 +77,14 @@ StdDev.SE <- function(data,
                                     se.method = mymethod,
                                     cleanOutliers=cleanOutliers,
                                     fitting.method=fitting.method, d.GLM.EN=d.GLM.EN,
-                                    freq.include=freq.include, freq.par=freq.par,,
+                                    freq.include=freq.include, freq.par=freq.par,
                                     ...)
       }
+
+      # Adding the correlations to the list
+      res <- Add_Correlations(res=res, data=data, cleanOutliers=cleanOutliers, corOut=corOut, IF.func=IF.SD, ...)
+
+      # Returning the output
       return(res)
     }
   }